ENH: Added verbose parameter to least_squares
