--- scipy/optimize/_lsq_bounds.py
@@ -1,162 +1,0 @@
-"""Utility functions to work with bound constraints."""
-
-import numpy as np
-
-
-def prepare_bounds(bounds, x0):
-    """Prepare bounds for usage in algorithms."""
-    lb, ub = [np.asarray(b, dtype=float) for b in bounds]
-    if lb.ndim == 0:
-        lb = np.resize(lb, x0.shape)
-
-    if ub.ndim == 0:
-        ub = np.resize(ub, x0.shape)
-
-    return lb, ub
-
-
-def in_bounds(x, lb, ub):
-    """Check if a point lies within bounds."""
-    return np.all((x >= lb) & (x <= ub))
-
-
-def step_size_to_bound(x, s, lb, ub):
-    """Compute a min_step size required to reach a bound.
-
-    The function computes a positive scalar t, such that x + s * t is on
-    the bound.
-
-    Returns
-    -------
-    step : float
-        Computed step. Non-negative value.
-    hits : ndarray of int with shape of x
-        Each element indicates whether a corresponding variable reaches the
-        bound:
-
-             *  0 - the bound was not hit.
-             * -1 - the lower bound was hit.
-             *  1 - the upper bound was hit.
-    """
-    non_zero = np.nonzero(s)
-    s_non_zero = s[non_zero]
-    steps = np.empty_like(x)
-    steps.fill(np.inf)
-    with np.errstate(over='ignore'):
-        steps[non_zero] = np.maximum((lb - x)[non_zero] / s_non_zero,
-                                     (ub - x)[non_zero] / s_non_zero)
-    min_step = np.min(steps)
-    return min_step, np.equal(steps, min_step) * np.sign(s).astype(int)
-
-
-def find_active_constraints(x, lb, ub, rtol=1e-10):
-    """Determine which constraints are active in a given point.
-
-    The threshold is computed using `rtol` and the absolute value of the
-    closest bound.
-
-    Returns
-    ------
-    active : ndarray of int with shape of x
-        Each component shows whether the corresponding constraint is active:
-
-             *  0 - a constraint is not active.
-             * -1 - a lower bound is active.
-             *  1 - a upper bound is active.
-    """
-    active = np.zeros_like(x, dtype=int)
-
-    if rtol == 0:
-        active[x <= lb] = -1
-        active[x >= ub] = 1
-        return active
-
-    lower_dist = x - lb
-    upper_dist = ub - x
-
-    lower_threshold = rtol * np.maximum(1, np.abs(lb))
-    upper_threshold = rtol * np.maximum(1, np.abs(ub))
-
-    lower_active = (np.isfinite(lb) &
-                    (lower_dist <= np.minimum(upper_dist, lower_threshold)))
-    active[lower_active] = -1
-
-    upper_active = (np.isfinite(ub) &
-                    (upper_dist <= np.minimum(lower_dist, upper_threshold)))
-    active[upper_active] = 1
-
-    return active
-
-
-def make_strictly_feasible(x, lb, ub, rstep=1e-10):
-    """Shift a point to the interior of a feasible region.
-
-    Each element of the returned vector is at least at a relative distance
-    `rstep` from the closest bound. If ``rstep=0`` then `np.nextafter` is used.
-    """
-    x_new = x.copy()
-
-    active = find_active_constraints(x, lb, ub, rstep)
-    lower_mask = np.equal(active, -1)
-    upper_mask = np.equal(active, 1)
-
-    if rstep == 0:
-        x_new[lower_mask] = np.nextafter(lb[lower_mask], ub[lower_mask])
-        x_new[upper_mask] = np.nextafter(ub[upper_mask], lb[upper_mask])
-    else:
-        x_new[lower_mask] = (lb[lower_mask] +
-                             rstep * np.maximum(1, np.abs(lb[lower_mask])))
-        x_new[upper_mask] = (ub[upper_mask] -
-                             rstep * np.maximum(1, np.abs(ub[upper_mask])))
-
-    tight_bounds = (x_new < lb) | (x_new > ub)
-    x_new[tight_bounds] = 0.5 * (lb[tight_bounds] + ub[tight_bounds])
-
-    return x_new
-
-
-def scaling_vector(x, g, lb, ub):
-    """Compute a scaling vector and its derivatives as described in papers
-    of Coleman and Li [1]_.
-
-    Components of a vector v are defined as follows:
-    ::
-
-               | ub[i] - x[i], if g[i] < 0 and ub[i] < np.inf
-        v[i] = | x[i] - lb[i], if g[i] > 0 and lb[i] > -np.inf
-               | 1,           otherwise
-
-    According to this definition v[i] >= 0 for all i. It differs from the
-    definition in paper [1]_ (eq. (2.2)), where the absolute value of v is
-    used. Both definitions are equivalent down the line.
-
-    Derivatives of v with respect to x take value 1, -1 or 0 depending on a
-    case.
-
-    Returns
-    -------
-    v : ndarray with shape of x
-        Scaling vector.
-    dv : ndarray with shape of x
-        Derivatives of v[i] with respect to x[i], diagonal elements of v's
-        Jacobian.
-
-    References
-    ----------
-    .. [1] Branch, M.A., T.F. Coleman, and Y. Li, "A Subspace, Interior,
-           and Conjugate Gradient Method for Large-Scale Bound-Constrained
-           Minimization Problems," SIAM Journal on Scientific Computing,
-           Vol. 21, Number 1, pp 1-23, 1999.
-    """
-    v = np.ones_like(x)
-    dv = np.zeros_like(x)
-
-    mask = (g < 0) & np.isfinite(ub)
-    v[mask] = ub[mask] - x[mask]
-    dv[mask] = -1
-
-    mask = (g > 0) & np.isfinite(lb)
-    v[mask] = x[mask] - lb[mask]
-    dv[mask] = 1
-
-    return v, dv
